Refactor HTTP handler to use new pagination logic for retrieving products by domain name. Update service and repository methods accordingly.

+++ internal/adapters/primary/http/response.go
@@ -12,7 +12,7 @@ type Response struct {
 type Pagination struct {
 	Page       int    `json:"page"`
 	PageSize   int    `json:"page_size"`
-	TotalItems int64  `json:"total_items"`
+	TotalItems int    `json:"total_items"`
 	TotalPages int    `json:"total_pages"`
 	NextPage   string `json:"next_page,omitempty"`
 	PrevPage   string `json:"prev_page,omitempty"`
